+++ processor/src/test/fixtures/input/com/example/inject/autofragment/MyFrameworkModelImpl.java
@@ -1,0 +1,6 @@
+package com.example.inject.autofragment;
+
+class MyFrameworkModelImpl
+  extends MyFrameworkModel
+{
+}