+++ processor/src/test/fixtures/input/com/example/inject/autofragment/OtherModel.java
@@ -1,0 +1,8 @@
+package com.example.inject.autofragment;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowEmpty = true )
+public abstract class OtherModel
+{
+}